backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic
gitpython==3.1.42
PyGithub==2.3.0
requests==2.31.0
google-generativeai
python-dotenv==1.0.1
langchain
langchain-core
langchain-google-genai
